test: 각 타입에 맞게 Unit Tests 파일을 분리
